Fix NSC-128 Add Support for OOREDOO EMS auto provisioning (script 2)
We need to change the current param format from json string to no
format string. Until now we set current as {_id:...,type:...,parent:...},
but that format fights with toml parser, so we shall set the current
param on toml file with _id value. So if _id 1234 current on toml should
has the same value. Type and parent are hardcoded on server-request with
admin and null values respectively.